+++ account_payment_return_import_iso20022/models/__init__.py
@@ -1,0 +1,1 @@
+from . import payment_return
