+++ account_payment_return_import_iso20022/readme/CONTRIBUTORS.rst
@@ -1,0 +1,5 @@
+* Carlos Dauden <pedro.baeza@tecnativa.com>
+* Pedro M. Baeza <pedro.baeza@tecnativa.com>
+* David Vidal <david.vidal@tecnativa.com>
+* Luis M. Ontalba <luis.martinez@tecnativa.com>
+* Thomas Binsfeld <thomas.binsfeld@acsone.eu>
