+++ account_payment_return_import_iso20022/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Module to import payment returns from pain.002.001.03 and camt.054.001.02 files.
