+++ account_payment_return_import_iso20022/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_import_iso20022
