--- account_payment_return_import_sepa_camt/__init__.py
@@ -1,1 +1,0 @@
-from . import wizards
