--- account_payment_return_import_sepa_camt/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Thomas Binsfeld <thomas.binsfeld@acsone.eu>
