--- account_payment_return_import_sepa_camt/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-Module to import payment returns from CAMT Bank to Customer Debit Credit Notification (camt.054.001.02).
