--- account_payment_return_import_sepa_camt/test_files/test-sepa-camt-unpaid.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Document xmlns="urn:iso:std:iso:20022:tech:xsd:camt.054.001.02" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<BkToCstmrDbtCdtNtfctn>
-		<GrpHdr>
-			<MsgId>MSGID99345678912</MsgId>
-			<CreDtTm>2016-10-08T20:49:00</CreDtTm>
-		</GrpHdr>
-		<Ntfctn>
-			<Id>ZY08XXXXXXIS634C</Id>
-			<CreDtTm>2016-10-08T20:49:00</CreDtTm>
-			<Acct>
-				<Id>
-					<IBAN>NL77ABNA0574908765</IBAN>
-				</Id>
-				<Ccy>EUR</Ccy>
-				<Svcr>
-					<FinInstnId>
-						<BIC>BOFIIE2D</BIC>
-					</FinInstnId>
-				</Svcr>
-			</Acct>
-			<TxsSummry>
-				<TtlCdtNtries>
-					<NbOfNtries>0</NbOfNtries>
-					<Sum>0.00</Sum>
-				</TtlCdtNtries>
-				<TtlDbtNtries>
-					<NbOfNtries>1</NbOfNtries>
-					<Sum>100</Sum>
-				</TtlDbtNtries>
-			</TxsSummry>
-			<Ntry>
-				<Amt Ccy="EUR">100.00</Amt>
-				<CdtDbtInd>DBIT</CdtDbtInd>
-				<Sts>BOOK</Sts>
-				<BookgDt>
-					<Dt>2016-10-08</Dt>
-				</BookgDt>
-				<BkTxCd>
-					<Domn>
-						<Cd>PMNT</Cd>
-						<Fmly>
-							<Cd>IDDT</Cd>
-							<SubFmlyCd>UPDD</SubFmlyCd>
-						</Fmly>
-					</Domn>
-					<Prtry>
-						<Cd>B3/0811</Cd>
-					</Prtry>
-				</BkTxCd>
-				<NtryDtls>
-					<TxDtls>
-						<Refs>
-							<InstrId>FA031615104</InstrId>
-							<EndToEndId>E2EID1</EndToEndId>
-							<MndtId>MANDATEID2</MndtId>
-						</Refs>
-						<RltdPties>
-							<Dbtr>
-								<Nm>DEBTOR4</Nm>
-							</Dbtr>
-							<DbtrAcct>
-								<Id>
-									<IBAN>IE16AIBK93110187654321</IBAN>
-								</Id>
-							</DbtrAcct>
-							<Cdtr>
-								<Nm>CREDITOR</Nm>
-								<Id>
-									<PrvtId>
-										<Othr>
-											<Id>IE03ZZZ123456</Id>
-											<SchmeNm>
-												<Prtry>SEPA</Prtry>
-											</SchmeNm>
-										</Othr>
-									</PrvtId>
-								</Id>
-							</Cdtr>
-						</RltdPties>
-						<RmtInf>
-							<Ustrd>FACTURE DU 15.03.16</Ustrd>
-						</RmtInf>
-						<RltdDts>
-							<IntrBkSttlmDt>2016-10-09</IntrBkSttlmDt>
-						</RltdDts>
-						<RtrInf>
-							<Orgtr>
-								<Id>
-									<OrgId>
-										<BICOrBEI>BOFIIE2D</BICOrBEI>
-									</OrgId>
-								</Id>
-							</Orgtr>
-							<Rsn>
-								<Cd>AC06</Cd>
-							</Rsn>
-							<AddtlInf>/RTYP/RTRN</AddtlInf>
-						</RtrInf>
-						<AddtlTxInf>/SQTP/RCUR/DTOFSGNT/2013-11-14</AddtlTxInf>
-					</TxDtls>
-				</NtryDtls>
-			</Ntry>
-		</Ntfctn>
-	</BkToCstmrDbtCdtNtfctn>
-</Document>
