--- account_payment_return_import_sepa_camt/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_import_sepa_camt
