--- account_payment_return_import_sepa_pain/__init__.py
@@ -1,3 +1,0 @@
-
-from . import models
-from . import wizard
