--- account_payment_return_import_sepa_pain/models/__init__.py
@@ -1,2 +1,0 @@
-
-from . import payment_return
