--- account_payment_return_import_sepa_pain/readme/CONTRIBUTORS.rst
@@ -1,4 +1,0 @@
-* Carlos Dauden <pedro.baeza@tecnativa.com>
-* Pedro M. Baeza <pedro.baeza@tecnativa.com>
-* David Vidal <david.vidal@tecnativa.com>
-* Luis M. Ontalba <luis.martinez@tecnativa.com>
