--- account_payment_return_import_sepa_pain/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-Module to import SEPA Direct Debit Unpaid Report File Format PAIN.002.001.03.
