--- account_payment_return_import_sepa_pain/tests/__init__.py
@@ -1,2 +1,0 @@
-
-from . import test_import_sepa_pain
