+++ setup/account_payment_return_import_iso20022/odoo/addons/account_payment_return_import_iso20022
@@ -1,0 +1,1 @@
+../../../../account_payment_return_import_iso20022
