--- setup/account_payment_return_import_sepa_camt/odoo/addons/account_payment_return_import_sepa_camt
@@ -1,1 +1,0 @@
-../../../../account_payment_return_import_sepa_camt
