--- setup/account_payment_return_import_sepa_pain/odoo/addons/account_payment_return_import_sepa_pain
@@ -1,1 +1,0 @@
-../../../../account_payment_return_import_sepa_pain
